refactor(tui): simplify sync test via shared fixture, single-pass message walk
Test review found the new reproducer was 175 lines that mostly
duplicated the existing sync.test.tsx scaffolding. Extract mount,
createFetch, wait, Probe, and worktree/directory constants into
a shared sync-fixture.tsx so both tests stay light and the
override hook lets each test customize only what it needs.

Also collapse the .map + for in sync.session.sync into a single
pass — saves an allocation and a traversal on the hot path.

No behavior change to the fix itself.
