+++ packages/opencode/src/cli/cmd/tui/context/sync.tsx
@@ -526,11 +526,12 @@ export const { use: useSync, provider: SyncProvider } = createSimpleContext({
               if (match.found) draft.session[match.index] = session.data!
               if (!match.found) draft.session.splice(match.index, 0, session.data!)
               draft.todo[sessionID] = todo.data ?? []
-              const messageData = messages.data ?? []
-              draft.message[sessionID] = messageData.map((x) => x.info)
-              for (const message of messageData) {
+              const infos: typeof draft.message[string] = []
+              for (const message of messages.data ?? []) {
+                infos.push(message.info)
                 draft.part[message.info.id] = message.parts
               }
+              draft.message[sessionID] = infos
               draft.session_diff[sessionID] = diff.data ?? []
             }),
           )
