+++ packages/opencode/test/cli/cmd/tui/sync-undefined-messages.test.tsx
@@ -2,156 +2,16 @@
 /**
  * Reproducer for #26560 — TUI crashes with
  *   `TypeError: undefined is not an object (evaluating 'f.data.map')`
- * when entering certain sessions on v1.14.42+.
- *
- * Root cause: `packages/opencode/src/cli/cmd/tui/context/sync.tsx` uses a
- * non-null assertion on `messages.data` while the same function defensively
- * guards `todo.data` and `diff.data` with `?? []`. When the messages endpoint
- * returns a non-2xx (e.g. the v1.14.42 validation-error envelope, or any
- * transient 4xx/5xx) the SDK leaves `data` undefined and the `.map` call
- * crashes the TUI before the user can see anything.
+ * when entering a session whose messages endpoint returns a non-2xx.
+ * The failure path is `sync.tsx#sync.session.sync` reading
+ * `messages.data!` while the SDK leaves `data` undefined on error.
  */
 import { describe, expect, test } from "bun:test"
-import { testRender } from "@opentui/solid"
-import { onMount } from "solid-js"
 import { Global } from "@opencode-ai/core/global"
-import { ArgsProvider } from "../../../../src/cli/cmd/tui/context/args"
-import { ExitProvider } from "../../../../src/cli/cmd/tui/context/exit"
-import { KVProvider } from "../../../../src/cli/cmd/tui/context/kv"
-import { ProjectProvider } from "../../../../src/cli/cmd/tui/context/project"
-import { SDKProvider, type EventSource } from "../../../../src/cli/cmd/tui/context/sdk"
-import { SyncProvider, useSync } from "../../../../src/cli/cmd/tui/context/sync"
 import { tmpdir } from "../../../fixture/fixture"
+import { directory, json, mount } from "./sync-fixture"
 
-const worktree = "/tmp/opencode"
-const directory = `${worktree}/packages/opencode`
 const sessionID = "ses_undef"
-
-function json(data: unknown, init?: ResponseInit) {
-  return new Response(JSON.stringify(data), {
-    ...init,
-    headers: { "content-type": "application/json", ...(init?.headers ?? {}) },
-  })
-}
-
-function eventSource(): EventSource {
-  return { subscribe: async () => () => {} }
-}
-
-function createFetch() {
-  const fetch = (async (input: RequestInfo | URL) => {
-    const url = new URL(input instanceof Request ? input.url : String(input))
-
-    switch (url.pathname) {
-      case "/agent":
-      case "/command":
-      case "/experimental/workspace":
-      case "/experimental/workspace/status":
-      case "/formatter":
-      case "/lsp":
-        return json([])
-      case "/config":
-      case "/experimental/resource":
-      case "/mcp":
-      case "/provider/auth":
-      case "/session/status":
-        return json({})
-      case "/config/providers":
-        return json({ providers: {}, default: {} })
-      case "/experimental/console":
-        return json({ consoleManagedProviders: [], switchableOrgCount: 0 })
-      case "/path":
-        return json({ home: "", state: "", config: "", worktree, directory })
-      case "/project/current":
-        return json({ id: "proj_test" })
-      case "/provider":
-        return json({ all: [], default: {}, connected: [] })
-      case "/session":
-        return json([
-          {
-            id: sessionID,
-            title: "broken",
-            time: { created: 0, updated: 0 },
-            version: "1.14.42",
-            directory,
-            project_id: "proj_test",
-          },
-        ])
-      case "/vcs":
-        return json({ branch: "main" })
-      case `/session/${sessionID}`:
-        return json({
-          id: sessionID,
-          title: "broken",
-          time: { created: 0, updated: 0 },
-          version: "1.14.42",
-          directory,
-          project_id: "proj_test",
-        })
-      case `/session/${sessionID}/messages`:
-        // The exact failure mode from #26560: server returns a non-2xx,
-        // SDK client surfaces `{ data: undefined, ... }`, the TUI's
-        // unguarded `messages.data!.map(...)` crashes the whole view.
-        return json({}, { status: 500 })
-      case `/session/${sessionID}/todo`:
-        return json([])
-      case `/session/${sessionID}/diff`:
-        return json([])
-    }
-
-    throw new Error(`unexpected request: ${url.pathname}`)
-  }) as typeof globalThis.fetch
-
-  return { fetch }
-}
-
-async function mount() {
-  const calls = createFetch()
-  let sync!: ReturnType<typeof useSync>
-  let done!: () => void
-  const ready = new Promise<void>((resolve) => {
-    done = resolve
-  })
-
-  const app = await testRender(() => (
-    <ArgsProvider>
-      <ExitProvider>
-        <KVProvider>
-          <SDKProvider url="http://test" directory={directory} fetch={calls.fetch} events={eventSource()}>
-            <ProjectProvider>
-              <SyncProvider>
-                <Probe
-                  onReady={(ctx) => {
-                    sync = ctx.sync
-                    done()
-                  }}
-                />
-              </SyncProvider>
-            </ProjectProvider>
-          </SDKProvider>
-        </KVProvider>
-      </ExitProvider>
-    </ArgsProvider>
-  ))
-
-  await ready
-  await waitFor(() => sync.status === "complete")
-  return { app, sync }
-}
-
-function Probe(props: { onReady: (ctx: { sync: ReturnType<typeof useSync> }) => void }) {
-  const sync = useSync()
-  onMount(() => props.onReady({ sync }))
-  return <box />
-}
-
-async function waitFor(predicate: () => boolean, timeout = 2000) {
-  const start = Date.now()
-  while (!predicate()) {
-    if (Date.now() - start > timeout) throw new Error("timed out waiting for condition")
-    await Bun.sleep(10)
-  }
-}
 
 describe("tui sync (#26560)", () => {
   test("entering a session whose messages endpoint errors does not crash sync", async () => {
@@ -160,12 +20,24 @@ describe("tui sync (#26560)", () => {
     Global.Path.state = tmp.path
     await Bun.write(`${tmp.path}/kv.json`, "{}")
 
-    const { app, sync } = await mount()
+    const sessionPayload = {
+      id: sessionID,
+      title: "broken",
+      time: { created: 0, updated: 0 },
+      version: "1.14.42",
+      directory,
+      project_id: "proj_test",
+    }
+    const { app, sync } = await mount((url) => {
+      if (url.pathname === `/session/${sessionID}`) return json(sessionPayload)
+      if (url.pathname === `/session/${sessionID}/messages`) return json({}, { status: 500 })
+      if (url.pathname === `/session/${sessionID}/todo`) return json([])
+      if (url.pathname === `/session/${sessionID}/diff`) return json([])
+      if (url.pathname === "/session") return json([sessionPayload])
+      return undefined
+    })
 
     try {
-      // Pre-fix this throws TypeError: undefined is not an object
-      // (evaluating 'f.data.map'). Post-fix the call resolves and the
-      // session simply has no messages locally cached.
       await expect(sync.session.sync(sessionID)).resolves.toBeUndefined()
     } finally {
       app.renderer.destroy()
